Create tests for #authenticate; Protect against JSON parsing errors
Previously, the function assumed that all responses coming from
SalesForce had valid JSON in the bodies. This caused unrecoverable
errors in production. Parsing of the JSON in this function is now
handled safely and will return with an error and the original body in
the event that the JSON is invalid.